Help and Back button
In this update I added a help room that we can put tutorial text in and I made a universal back button. To get that to work I tweek the parent code for buttons to make a gobal variable that will note the room you are in right before you leave using the button. The back button then calls that variable to return you where you were!