Added room jumping keys and end room
Added 9 and 0 as keys to go to previous room or next room (will throw an error if there are no previous or next rooms to go to). I also put the rooms in a sort of order with mine the first lvl after the menu and Alex White 2 and Alex Whit 3 since with those birbs its pretty hard to beat. Took out my testing room. Added a end room.